verification of codes
