+++ src/actions/Main.js
@@ -3,10 +3,6 @@ export const GET_CODE_STATUS = 'GET_CODE_STATUS';
 export const APP_LOADED = 'APP_LOADED';
 export const RESIZE_WINDOW = 'RESIZE_WINDOW';
 
-export const GET_SIGNUP_CODESTATUS = 'GET_SIGNUP_CODESTATUS';
-export const GET_SIGNUP_CODESTATUS_SUCCESS = 'GET_SIGNUP_CODESTATUS_SUCCESS';
-export const GET_SIGNUP_CODESTATUS_FAIL = 'GET_SIGNUP_CODESTATUS_FAIL';
-
 export const GET_SIGNUP_CONFIG = 'GET_SIGNUP_CONFIG';
 export const GET_SIGNUP_CONFIG_SUCCESS = 'GET_SIGNUP_CONFIG_SUCCESS';
 export const GET_SIGNUP_CONFIG_FAIL = 'GET_SIGNUP_CONFIG_FAIL';
@@ -36,18 +32,6 @@ export function resizeWindow () {
             window_size: getWindowSize()
         }
     };
-}
-
-
-export function getCodeStatusFail (err) {
-  return {
-    type: GET_SIGNUP_CODESTATUS_FAIL,
-    error: true,
-    payload: {
-      error: err,
-      message: err
-    }
-  };
 }
 
 export function getConfig () {
