+++ src/containers/CodeVerified.js
@@ -7,13 +7,14 @@ import CodeVerified from 'components/CodeVerified';
 const mapStateToProps = (state, props) => {
     const url = '';
     return {
-        dashboard_url: url,
+        dashboard_url: state.verified.dashboard_url,
         password: state.verified.password,
         eppn: state.verified.eppn,
         nonce: state.verified.nonce,
         ts: state.verified.timestamp,
-        token: state.verified.token,
-        email: state.email.email
+        token: state.verified.auth_token,
+        email: state.verified.email,
+        status: state.verified.status
     }
 };
