+++ src/store.js
@@ -7,6 +7,7 @@ import { routerReducer } from 'react-router-redux';
 import captchaReducer from 'reducers/Captcha';
 import mainReducer from 'reducers/Main';
 import emailReducer from 'reducers/Email';
+import verifiedReducer from "reducers/CodeVerified";
 import notificationsReducer from 'reducers/Notifications';
 
 const eduIDApp = combineReducers({
